Agrego clases varias

+++ dominio/Producto.cs
@@ -12,12 +12,12 @@ namespace dominio
         public int IdProducto { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
-        public float Precio { get; set; }
+        public decimal Precio { get; set; }
         public int Stock { get; set; }
         public Categoria Categoria { get; set; }
         public Marca Marca { get; set; }
         public Proveedor Proveedor { get; set; }
-        public float Peso { get; set; }
+        public decimal Peso { get; set; }
         public bool Estado { get; set; }
         public List<Imagen> ListImagen { get; set; }
